E2e Daemonset
Update e2e script to include optional flag (-d|--daemonset) to deploy
the DP as a daemonset rather than running the binary directly on the
host.

Update CI to use same. CI now runs once normally and once in dameonset.
Reason for this is the DP logs are not printed to screen during a
daemonset run.