Update e2e test app to use /pkg/uds
e2e/udsTest.go generally tidied up and improved code quality. Now uses
our /pkg/uds which is much cleaner than the code that had been created
and badly altered over time as requirements changed.

/pkg/uds itself needed some updates to make it more portable and usable
to code other than /pkg/cndp and this resulted in necessary updates to
/pkg/cndp and /pkg/uds/fake.

Due to dependency on /pkg/uds, the test app is no longer built in the
e2e Dockerfile, but in the e2e script before the image is created.
Changes in Dockerfile and e2e script reflect this.