Netdev functions (#50)
* Added functionality to obtain netdev from a MAC Address
* Added GetDeviceByPCI functionality to networking package
GetDeviceByPCI takes in a pci address as a string and returns a netlink Link to the device.
GetDeviceByMAC is similar, but takes in a mac address instead. Proper logging statements have been added.
* Adjusted return output from netdev functions
GetDeviceByMAC and GetDeviceByPCI now return the device name as a string instead of a netlink Link as per the ask.