GRPC update (#69)
Bug Fix for K8s v25
I have replicated this error on the newest version of kubernetes v25.3 with CRI-O runtime.
The error is due to the order of initializing grpc with kubelet. Newest version of kubelet requires grpc to first start and then register a session.
I've also tested the reordering of the grpc functions with older version of k8s (22.5) which works as expected.